requirements.txt
# EagleEye Lite - Dependencies
# AI/ML
openai>=1.0.0
sentence-transformers>=2.2.0

# PDF Processing
pdfplumber>=0.10.0
easyocr>=1.7.0
pdf2image>=1.16.0
Pillow>=10.0.0

# Vector Database
chromadb>=0.4.0

# LangGraph/LangChain
langgraph>=0.1.0
langchain-core>=0.2.0

# Data Validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.0.0
